Write error output to stdout in Wasm
